// src/app/services/api/model/cd.ts
/**
 * Snippets API
 * Test description
 *
 * OpenAPI spec version: v1
 * Contact: contact@snippets.local
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface CD {
    id?: number;
    name: string;
    artist: string;
    record_company: string;
    genre: string;
    ean_code: string;
    price: string;
    price_currency?: string;
    published_by?: number;
    user?: string;
    created_at?: Date;
    updated_at?: Date;
}